chore: build search component 🎢
